+++ libs/src/sdk/index.ts
@@ -29,4 +29,3 @@ export * from './services'
 export * from './config'
 export { ParseRequestError } from './utils/parseRequestParameters'
 export { RendezvousHash } from '../utils'
-export * as utilities from '../utils'
